add new filter with its endpoints for most purchases and fix pagination on all admin dashboard, also change it to a tab style one
