+++ backend/api/httpx/form/purchase.go
@@ -43,7 +43,13 @@ type Purchases struct {
 }
 
 type User struct {
+	ID    string `json:"id"`
 	Name  string `json:"name"`
 	Email string `json:"email"`
 	Phone string `json:"phone"`
+}
+
+type MostPurchases struct {
+	User     User `json:"user"`
+	Quantity int  `json:"quantity"`
 }
