+++ backend/internal/repository/purchase_repository.go
@@ -15,8 +15,12 @@ type PurchaseRepository interface {
 	GetAll(
 		ctx context.Context,
 		filters dto.GetAllPurchases,
-	) ([]form.Purchases, error)
+	) ([]form.Purchases, int, error)
 	UpdateStatus(ctx context.Context, purchaseID, status string) error
+	GetLeaderboard(
+		ctx context.Context,
+		filters dto.GetMostPurchases,
+	) ([]form.MostPurchases, error)
 }
 
 type purchaseRepo struct{ db database.DB }
@@ -48,16 +52,17 @@ func (r *purchaseRepo) Create(ctx context.Context, p *types.Purchase) error {
 func (r *purchaseRepo) GetAll(
 	ctx context.Context,
 	filters dto.GetAllPurchases,
-) ([]form.Purchases, error) {
+) ([]form.Purchases, int, error) {
 	var args []interface{}
-	query := `SELECT u.name, u.email, u.phone,
+	query := `SELECT u.id, u.name, u.email, u.phone,
     p.id, p.quantity, p.monto_bs, p.monto_usd, p.payment_method,
-    p.transaction_digits, p.payment_screenshot, p.status, p.created_at
+    p.transaction_digits, p.payment_screenshot, p.status, p.created_at,
+	COUNT(*) OVER() as total_count
 	FROM purchases p JOIN users u ON p.user_id = u.id `
 
 	argIdx := 1
 	if filters.PurchaseStatus != "" {
-		query += "WHERE P.status = $" + fmt.Sprint(argIdx) + " "
+		query += "WHERE p.status = $" + fmt.Sprint(argIdx) + " "
 		args = append(args, filters.PurchaseStatus)
 		argIdx++
 	}
@@ -78,14 +83,17 @@ func (r *purchaseRepo) GetAll(
 
 	rows, err := r.db.Query(ctx, query, args...)
 	if err != nil {
-		return nil, err
+		return nil, 0, err
 	}
 	defer rows.Close()
 
 	var purchases []form.Purchases
+	var total int
 	for rows.Next() {
 		var p form.Purchases
+		var rowTotal int
 		err := rows.Scan(
+			&p.User.ID,
 			&p.User.Name,
 			&p.User.Email,
 			&p.User.Phone,
@@ -98,15 +106,18 @@ func (r *purchaseRepo) GetAll(
 			&p.PaymentScreenshot,
 			&p.Status,
 			&p.CreatedAt,
+			&rowTotal,
 		)
 		if err != nil {
-			return nil, err
+			return nil, 0, err
 		}
-
+		if total == 0 {
+			total = rowTotal
+		}
 		purchases = append(purchases, p)
 	}
 
-	return purchases, nil
+	return purchases, total, nil
 }
 
 func (r *purchaseRepo) UpdateStatus(
@@ -119,4 +130,56 @@ func (r *purchaseRepo) UpdateStatus(
 		status, purchaseID,
 	)
 	return err
+}
+
+func (r *purchaseRepo) GetLeaderboard(
+	ctx context.Context,
+	filters dto.GetMostPurchases,
+) ([]form.MostPurchases, error) {
+	var args []interface{}
+	query := `SELECT u.id, u.name, u.email, u.phone,
+    	SUM(p.quantity) as quantity
+		FROM purchases p JOIN users u ON p.user_id = u.id
+		WHERE p.status = 'verified'
+		GROUP BY u.id, u.name, u.email, u.phone
+		ORDER BY quantity DESC `
+
+	argIdx := 1
+	// Pagination logic
+	perPage := filters.ItemCount
+	if perPage <= 0 {
+		perPage = 10
+	}
+	page := filters.Page
+	if page <= 0 {
+		page = 1
+	}
+	offset := (page - 1) * perPage
+	query += fmt.Sprintf("LIMIT $%d OFFSET $%d", argIdx, argIdx+1)
+	args = append(args, perPage, offset)
+
+	rows, err := r.db.Query(ctx, query, args...)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+
+	var leaderboard []form.MostPurchases
+	for rows.Next() {
+		var entry form.MostPurchases
+		err := rows.Scan(
+			&entry.User.ID,
+			&entry.User.Name,
+			&entry.User.Email,
+			&entry.User.Phone,
+			&entry.Quantity,
+		)
+		if err != nil {
+			return nil, err
+		}
+
+		leaderboard = append(leaderboard, entry)
+	}
+
+	return leaderboard, nil
 }
