+++ frontend/src/utils/arrays.ts
@@ -1,0 +1,3 @@
+export function safeArray<T>(data: unknown): T[] {
+  return Array.isArray(data) ? data : [];
+}
